feat: Add CopyToClipboardButton component for copying content to clipboard
A new component, CopyToClipboardButton, has been added to the project. This component allows users to easily copy content to their clipboard with a single click. It provides a button with the option to copy the specified content and displays a success message upon successful copying. This feature enhances the user experience and improves the usability of the application.