feat: Add thumbnail_url field to updateCourseAction
Add the thumbnail_url field to the updateCourseAction function in courseActions.ts. This allows for updating the thumbnail URL of a course. By including this field, we provide the ability to easily update the thumbnail image associated with a course, enhancing the visual representation of the course in the user interface.