feat: Refactor CreateCourseLessonPage to use sequence instead of exam_date for lesson data
Update the CreateCourseLessonPage component to use the sequence field instead of the exam_date field for lesson data. This change improves the clarity and consistency of the code by using a more appropriate field name. Additionally, the date field is formatted using the dayjs library to display it in the "DD/MM/YYYY" format.