refactor: Update form layout in forgot-password page
This commit updates the form layout in the forgot-password page by increasing the gap between form elements. The previous gap of 2 has been changed to 8 to provide better visual separation between the input fields. This change improves the user experience by making the form more readable and easier to navigate.